requirements.txt
attrs==20.3.0
boto3==1.17.44
botocore==1.20.44
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
linecache2==1.0.0
mangum==0.10.0
MarkupSafe==1.1.1
oslo.i18n==5.0.1
packaging==20.9
pbr==5.5.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
starlette==0.13.6
toml==0.10.2
traceback2==1.4.0
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
Werkzeug==1.0.1
wrapt==1.12.1
